Finish car module;

+++ src/cars/cars.service.ts
@@ -1,8 +1,9 @@
-import {BadRequestException, Injectable} from '@nestjs/common';
+import {BadRequestException, forwardRef, Inject, Injectable} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { Car } from './car.entity';
+import { RoutesService } from '../routes/routes.service';
 import { CarStatusesService } from '../car-statuses/car-statuses.service';
 import { TransportTypesService } from '../transport-types/transport-types.service';
 
@@ -13,6 +14,8 @@ export class CarsService {
         private carRepository: Repository<Car>,
         private carStatusesService: CarStatusesService,
         private transportTypesService: TransportTypesService,
+        @Inject(forwardRef(() => RoutesService))
+        private routesService: RoutesService,
     ) {}
 
     private async addRelations(car: Car): Promise<void> {
@@ -22,12 +25,17 @@ export class CarsService {
         car.transportType = transportType;
     }
 
-    async findAll(): Promise<Car[]> {
-        return this.carRepository.find({ relations: ['carStatus', 'transportType'] });
+    async findAll(params?): Promise<Car[]> {
+        return this.carRepository.find({
+            where: params,
+            relations: ['carStatus', 'transportType']
+        });
     }
 
-    async findCarById(id: number): Promise<Car> {
-        const car = await this.carRepository.findOne(id);
+    async findById(id: number): Promise<Car> {
+        const car = await this.carRepository.findOne(id, {
+            relations: ['carStatus', 'transportType']
+        });
         if (!car) {
             throw new BadRequestException('Car not found');
         }
@@ -40,15 +48,17 @@ export class CarsService {
     }
 
     async update(id: number, newCar: Car): Promise<Car> {
-        const car = await this.findCarById(id);
+        const car = await this.findById(id);
         await this.addRelations(newCar);
         const updatedCar = { ...car, ...newCar };
-        await this.carRepository.save(updatedCar);
-        return updatedCar;
+        return await this.carRepository.save(updatedCar);
     }
 
     async delete(id: number): Promise<void> {
-        await this.findCarById(id);
+        const car = await this.findById(id);
+        if (car.routes.length > 0) {
+            throw new BadRequestException('Car cannot be deleted while it relate with routes');
+        }
         await this.carRepository.delete(id);
     }
 }
